deployment_package/requirements.txt
Pillow
requests
boto3
numpy